+++ CHANGELOG.md
@@ -5,6 +5,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+## [8.5.0]
+### Added
+- Add Base network support ([#2310](https://github.com/MetaMask/test-dapp/pull/310))
+
+## [8.4.0]
+### Added
+- Add Malformed Transactions ([#295](https://github.com/MetaMask/test-dapp/pull/295))
+- Adds transferFromTokens error to token addresses text box ([#304](https://github.com/MetaMask/test-dapp/pull/304))
+- Add Increase Token Allowance btn ([#302](https://github.com/MetaMask/test-dapp/pull/302))
+
+## [8.3.0]
+### Fixed
+- Replace deprecated `window.ethereum.selectedAddress` property with local account value ([#300](https://github.com/MetaMask/test-dapp/pull/300))
+
 ## [8.2.0]
 ### Added
 - Add Wallet Connect support ([#284](https://github.com/MetaMask/test-dapp/pull/284))
@@ -147,7 +161,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix repository standardization issues ([#118](https://github.com/MetaMask/test-dapp/pull/118))
 - Fix addEthereumChain button disable logic ([#93](https://github.com/MetaMask/test-dapp/pull/93))
 
-[Unreleased]: https://github.com/MetaMask/test-dapp/compare/v8.2.0...HEAD
+[Unreleased]: https://github.com/MetaMask/test-dapp/compare/v8.5.0...HEAD
+[8.5.0]: https://github.com/MetaMask/test-dapp/compare/v8.4.0...v8.5.0
+[8.4.0]: https://github.com/MetaMask/test-dapp/compare/v8.3.0...v8.4.0
+[8.3.0]: https://github.com/MetaMask/test-dapp/compare/v8.2.0...v8.3.0
 [8.2.0]: https://github.com/MetaMask/test-dapp/compare/v8.1.0...v8.2.0
 [8.1.0]: https://github.com/MetaMask/test-dapp/compare/v8.0.0...v8.1.0
 [8.0.0]: https://github.com/MetaMask/test-dapp/compare/v7.3.1...v8.0.0